--- a/src/extension.ts
+++ b/src/extension.ts
@@ -7,6 +7,7 @@ import * as clusterCommands from './commands/clusterCommands';
 import * as topicCommands from './commands/topicCommands';
 import * as consumerGroupCommands from './commands/consumerGroupCommands';
 import * as brokerCommands from './commands/brokerCommands';
+import * as clusterDashboardCommands from './commands/clusterDashboardCommands';
 
 // Global client manager instance for cleanup on deactivation
 let clientManager: KafkaClientManager;
@@ -145,6 +146,12 @@ export async function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(
         vscode.commands.registerCommand('kafka.findBroker', async () => {
             await brokerCommands.findBroker(clientManager);
+        })
+    );
+
+    context.subscriptions.push(
+        vscode.commands.registerCommand('kafka.showClusterDashboard', async (node) => {
+            await clusterDashboardCommands.showClusterDashboard(clientManager, context, node);
         })
     );
 }